+++ app/main.development.js
@@ -3,25 +3,25 @@ import { concat, isString } from "lodash";
 import { initGlobalCfg, validateGlobalCfgFile, setMustOpenForm } from "./config";
 import { initWalletCfg, getWalletCfg, newWalletConfigCreation, readDcrdConfig, createTempDcrdConf } from "./config";
 import fs from "fs-extra";
-import os from "os";
 import path from "path";
 import parseArgs from "minimist";
 import stringArgv from "string-argv";
 import { appLocaleFromElectronLocale, default as locales } from "./i18n/locales";
-import { createLogger } from "./main_dev/logging";
+import { createLogger, AddToLog, lastLogLine } from "./main_dev/logging";
 import { Buffer } from "buffer";
-import { OPTIONS, USAGE_MESSAGE, VERSION_MESSAGE, MAX_LOG_LENGTH, BOTH_CONNECTION_ERR_MESSAGE } from "./main_dev/constants";
+import { OPTIONS, USAGE_MESSAGE, VERSION_MESSAGE, BOTH_CONNECTION_ERR_MESSAGE } from "./main_dev/constants";
 import { appDataDirectory, getDcrdPath, dcrctlCfg, dcrdCfg, getDefaultWalletFilesPath } from "./main_dev/paths";
 import { dcrwalletCfg, getWalletPath, getExecutablePath, getWalletsDirectoryPath, getWalletsDirectoryPathNetwork, getDefaultWalletDirectory } from "./main_dev/paths";
 import { getGlobalCfgPath, getDecreditonWalletDBPath, getWalletDBPathFromWallets, getDcrdRpcCert, getDirectoryLogs, checkAndInitWalletCfg } from "./main_dev/paths";
 import { installSessionHandlers, reloadAllowedExternalRequests, allowStakepoolRequests } from "./main_dev/externalRequests";
 import { setupProxy } from "./main_dev/proxy";
+import { cleanShutdown, launchDCRD, launchDCRWallet } from "./main_dev/launch"
 
 // setPath as decrediton
 app.setPath("userData", appDataDirectory());
 
 const argv = parseArgs(process.argv.slice(1), OPTIONS);
-let debug = argv.debug || process.env.NODE_ENV === "development";
+const debug = argv.debug || process.env.NODE_ENV === "development";
 const logger = createLogger(debug);
 
 // Verify that config.json is valid JSON before fetching it, because
@@ -98,61 +98,6 @@ process.on("uncaughtException", err => {
   logger.log("error", "UNCAUGHT EXCEPTION", err);
   throw err;
 });
-
-function closeDCRW() {
-  if (require("is-running")(dcrwPID) && os.platform() != "win32") {
-    logger.log("info", "Sending SIGINT to dcrwallet at pid:" + dcrwPID);
-    process.kill(dcrwPID, "SIGINT");
-    dcrwPID = null;
-  }
-}
-
-function closeDCRD() {
-  if (require("is-running")(dcrdPID) && os.platform() != "win32") {
-    logger.log("info", "Sending SIGINT to dcrd at pid:" + dcrdPID);
-    process.kill(dcrdPID, "SIGINT");
-  }
-}
-
-function closeClis() {
-  // shutdown daemon and wallet.
-  // Don't try to close if not running.
-  if(dcrdPID && dcrdPID !== -1)
-    closeDCRD();
-  if(dcrwPID && dcrwPID !== -1)
-    closeDCRW();
-}
-
-async function cleanShutdown() {
-  // Attempt a clean shutdown.
-  return new Promise(resolve => {
-    const cliShutDownPause = 2; // in seconds.
-    const shutDownPause = 3; // in seconds.
-    closeClis();
-    // Sent shutdown message again as we have seen it missed in the past if they
-    // are still running.
-    setTimeout(function () { closeClis(); }, cliShutDownPause * 1000);
-    logger.log("info", "Closing decrediton.");
-
-    let shutdownTimer = setInterval(function(){
-      const stillRunning = (require("is-running")(dcrdPID) && os.platform() != "win32");
-
-      if (!stillRunning) {
-        logger.log("info", "Final shutdown pause. Quitting app.");
-        clearInterval(shutdownTimer);
-        if (mainWindow) {
-          mainWindow.webContents.send("daemon-stopped");
-          setTimeout(() => {mainWindow.close(); app.quit();}, 1000);
-        } else {
-          app.quit();
-        }
-        resolve(true);
-      }
-      logger.log("info", "Daemon still running in final shutdown pause. Waiting.");
-
-    }, shutDownPause*1000);
-  });
-}
 
 const installExtensions = async () => {
   if (process.env.NODE_ENV === "development") {
@@ -277,7 +222,10 @@ ipcMain.on("start-wallet", (event, walletPath, testnet) => {
   }
   initWalletCfg(testnet, walletPath);
   try {
-    dcrwPID = launchDCRWallet(walletPath, testnet);
+    const pidAndPort = launchDCRWallet(mainWindow, dcrwalletLogs, daemonIsAdvanced, walletPath, testnet, logger);
+    dcrwPID = pidAndPort.dcrwPID;
+    dcrwPort = pidAndPort.dcrwPort;
+    logger.log("info", "this is PID: %s this is PORT: %s\n\n\n", dcrwPID, dcrwPort)
   } catch (e) {
     logger.log("error", "error launching dcrwallet: " + e);
   }
@@ -334,7 +282,7 @@ ipcMain.on("check-daemon", (event, rpcCreds, testnet) => {
 });
 
 ipcMain.on("clean-shutdown", async function(event){
-  const stopped = await cleanShutdown();
+  const stopped = await cleanShutdown(logger, mainWindow, app, dcrdPID, dcrwPID);
   event.sender.send("clean-shutdown-finished", stopped);
 });
 
@@ -362,12 +310,6 @@ ipcMain.on("get-decrediton-logs", (event) => {
   event.returnValue = "decrediton logs!";
 });
 
-function lastLogLine(log) {
-  let lastLineIdx = log.lastIndexOf(os.EOL, log.length - os.EOL.length -1);
-  let lastLineBuff = log.slice(lastLineIdx).toString("utf-8");
-  return lastLineBuff.trim();
-}
-
 ipcMain.on("get-last-log-line-dcrd", event => {
   event.returnValue = lastLogLine(dcrdLogs);
 });
@@ -384,220 +326,6 @@ ipcMain.on("set-previous-wallet", (event, cfg) => {
   previousWallet = cfg;
   event.returnValue = true;
 });
-
-const AddToLog = (destIO, destLogBuffer, data) => {
-  var dataBuffer = Buffer.from(data);
-  if (destLogBuffer.length + dataBuffer.length > MAX_LOG_LENGTH) {
-    destLogBuffer = destLogBuffer.slice(destLogBuffer.indexOf(os.EOL,dataBuffer.length)+1);
-  }
-  debug && destIO.write(data);
-  return Buffer.concat([ destLogBuffer, dataBuffer ]);
-};
-
-// DecodeDaemonIPCData decodes messages from an IPC message received from dcrd/
-// dcrwallet using their internal IPC protocol.
-// NOTE: very simple impl for the moment, will break if messages get split
-// between data calls.
-const DecodeDaemonIPCData = (data, cb) => {
-  let i = 0;
-  while (i < data.length) {
-    if (data[i++] !== 0x01) throw "Wrong protocol version when decoding IPC data";
-    const mtypelen = data[i++];
-    const mtype = data.slice(i, i+mtypelen).toString("utf-8");
-    i += mtypelen;
-    const psize = data.readUInt32LE(i);
-    i += 4;
-    const payload = data.slice(i, i+psize);
-    i += psize;
-    cb(mtype, payload);
-  }
-};
-
-const launchDCRD = (daemonPath, appdata, testnet) => {
-  var spawn = require("child_process").spawn;
-  let args = [];
-  let newConfig = {};
-  if(appdata){
-    args = [ `--appdata=${appdata}` ];
-    newConfig = readDcrdConfig(appdata, testnet);
-    newConfig.rpc_cert = getDcrdRpcCert(appdata);
-  } else {
-    args = [ `--configfile=${dcrdCfg(daemonPath)}` ];
-    newConfig = readDcrdConfig(daemonPath, testnet);
-    newConfig.rpc_cert = getDcrdRpcCert();
-  }
-  if (testnet) {
-    args.push("--testnet");
-  }
-  // Check to make sure that the rpcuser and rpcpass were set in the config
-  if (!newConfig.rpc_user || !newConfig.rpc_password) {
-    const errorMessage =  "No " + `${!newConfig.rpc_user ? "rpcuser " : "" }` + `${!newConfig.rpc_user && !newConfig.rpc_password ? "and " : "" }` + `${!newConfig.rpc_password ? "rpcpass " : "" }` + "set in " + `${appdata ? appdata : getDcrdPath()}` + "/dcrd.conf.  Please set them and restart.";
-    logger.log("error", errorMessage);
-    mainWindow.webContents.executeJavaScript("alert(\"" + `${errorMessage}` + "\");");
-    mainWindow.webContents.executeJavaScript("window.close();");
-  }
-
-  var dcrdExe = getExecutablePath("dcrd", argv.customBinPath);
-  if (!fs.existsSync(dcrdExe)) {
-    logger.log("error", "The dcrd file does not exists");
-    return;
-  }
-
-  if (os.platform() == "win32") {
-    try {
-      const util = require("util");
-      const win32ipc = require("./node_modules/win32ipc/build/Release/win32ipc.node");
-      var pipe = win32ipc.createPipe("out");
-      args.push(util.format("--piperx=%d", pipe.readEnd));
-    } catch (e) {
-      logger.log("error", "can't find proper module to launch dcrd: " + e);
-    }
-  }
-
-  logger.log("info", `Starting ${dcrdExe} with ${args}`);
-
-  var dcrd = spawn(dcrdExe, args, {
-    detached: os.platform() == "win32",
-    stdio: [ "ignore", "pipe", "pipe" ]
-  });
-
-  dcrd.on("error", function (err) {
-    logger.log("error", "Error running dcrd.  Check logs and restart! " + err);
-    mainWindow.webContents.executeJavaScript("alert(\"Error running dcrd.  Check logs and restart! " + err + "\");");
-    mainWindow.webContents.executeJavaScript("window.close();");
-  });
-
-  dcrd.on("close", (code) => {
-    if (daemonIsAdvanced)
-      return;
-    if (code !== 0) {
-      logger.log("error", "dcrd closed due to an error.  Check dcrd logs and contact support if the issue persists.");
-      mainWindow.webContents.executeJavaScript("alert(\"dcrd closed due to an error.  Check dcrd logs and contact support if the issue persists.\");");
-      mainWindow.webContents.executeJavaScript("window.close();");
-    } else {
-      logger.log("info", `dcrd exited with code ${code}`);
-    }
-  });
-
-  dcrd.stdout.on("data", (data) => dcrdLogs = AddToLog(process.stdout, dcrdLogs, data));
-  dcrd.stderr.on("data", (data) => dcrdLogs = AddToLog(process.stderr, dcrdLogs, data));
-
-  newConfig.pid = dcrd.pid;
-  logger.log("info", "dcrd started with pid:" + newConfig.pid);
-
-  dcrd.unref();
-  return newConfig;
-};
-
-const launchDCRWallet = (walletPath, testnet) => {
-  var spawn = require("child_process").spawn;
-  var args = [ "--configfile=" + dcrwalletCfg(getWalletPath(testnet, walletPath)) ];
-
-  const cfg = getWalletCfg(testnet, walletPath);
-
-  args.push("--ticketbuyer.nospreadticketpurchases");
-  args.push("--ticketbuyer.balancetomaintainabsolute=" + cfg.get("balancetomaintain"));
-  args.push("--ticketbuyer.maxfee=" + cfg.get("maxfee"));
-  args.push("--ticketbuyer.maxpricerelative=" + cfg.get("maxpricerelative"));
-  args.push("--ticketbuyer.maxpriceabsolute=" + cfg.get("maxpriceabsolute"));
-  args.push("--ticketbuyer.maxperblock=" + cfg.get("maxperblock"));
-  args.push("--addridxscanlen=" + cfg.get("gaplimit"));
-
-  var dcrwExe = getExecutablePath("dcrwallet", argv.customBinPath);
-  if (!fs.existsSync(dcrwExe)) {
-    logger.log("error", "The dcrwallet file does not exists");
-    return;
-  }
-
-  if (os.platform() == "win32") {
-    try {
-      const util = require("util");
-      const win32ipc = require("./node_modules/win32ipc/build/Release/win32ipc.node");
-      var pipe = win32ipc.createPipe("out");
-      args.push(util.format("--piperx=%d", pipe.readEnd));
-    } catch (e) {
-      logger.log("error", "can't find proper module to launch dcrwallet: " + e);
-    }
-  } else {
-    args.push("--rpclistenerevents");
-    args.push("--pipetx=4");
-  }
-
-  // Add any extra args if defined.
-  if (argv.extrawalletargs !== undefined && isString(argv.extrawalletargs)) {
-    args = concat(args, stringArgv(argv.extrawalletargs));
-  }
-
-  logger.log("info", `Starting ${dcrwExe} with ${args}`);
-
-  var dcrwallet = spawn(dcrwExe, args, {
-    detached: os.platform() == "win32",
-    stdio: [ "ignore", "pipe", "pipe", "ignore", "pipe" ]
-  });
-
-  const notifyGrpcPort = (port) => {
-    dcrwPort = port;
-    logger.log("info", "wallet grpc running on port", port);
-    mainWindow.webContents.send("dcrwallet-port", port);
-  };
-
-  dcrwallet.stdio[4].on("data", (data) => DecodeDaemonIPCData(data, (mtype, payload) => {
-    if (mtype === "grpclistener") {
-      const intf = payload.toString("utf-8");
-      const matches = intf.match(/^.+:(\d+)$/);
-      if (matches) {
-        notifyGrpcPort(matches[1]);
-      } else {
-        logger.log("error", "GRPC port not found on IPC channel to dcrwallet: " + intf);
-      }
-    }
-  }));
-
-  dcrwallet.on("error", function (err) {
-    logger.log("error", "Error running dcrwallet.  Check logs and restart! " + err);
-    mainWindow.webContents.executeJavaScript("alert(\"Error running dcrwallet.  Check logs and restart! " + err + "\");");
-    mainWindow.webContents.executeJavaScript("window.close();");
-  });
-
-  dcrwallet.on("close", (code) => {
-    if(daemonIsAdvanced)
-      return;
-    if (code !== 0) {
-      logger.log("error", "dcrwallet closed due to an error.  Check dcrwallet logs and contact support if the issue persists.");
-      mainWindow.webContents.executeJavaScript("alert(\"dcrwallet closed due to an error.  Check dcrwallet logs and contact support if the issue persists.\");");
-      mainWindow.webContents.executeJavaScript("window.close();");
-    } else {
-      logger.log("info", `dcrwallet exited with code ${code}`);
-    }
-  });
-
-  const addStdoutToLogListener = (data) => dcrwalletLogs = AddToLog(process.stdout, dcrwalletLogs, data);
-
-  // waitForGrpcPortListener is added as a stdout on("data") listener only on
-  // win32 because so far that's the only way we found to get back the grpc port
-  // on that platform. For linux/macOS users, the --pipetx argument is used to
-  // provide a pipe back to decrediton, which reads the grpc port in a secure and
-  // reliable way.
-  const waitForGrpcPortListener = (data) => {
-    const matches = /DCRW: gRPC server listening on [^ ]+:(\d+)/.exec(data);
-    if (matches) {
-      notifyGrpcPort(matches[1]);
-      // swap the listener since we don't need to keep looking for the port
-      dcrwallet.stdout.removeListener("data", waitForGrpcPortListener);
-      dcrwallet.stdout.on("data", addStdoutToLogListener);
-    }
-    dcrwalletLogs = AddToLog(process.stdout, dcrwalletLogs, data);
-  };
-
-  dcrwallet.stdout.on("data", os.platform() == "win32" ? waitForGrpcPortListener : addStdoutToLogListener);
-  dcrwallet.stderr.on("data", (data) => dcrwalletLogs = AddToLog(process.stderr, dcrwalletLogs, data));
-
-  dcrwPID = dcrwallet.pid;
-  logger.log("info", "dcrwallet started with pid:" + dcrwPID);
-
-  dcrwallet.unref();
-  return dcrwPID;
-};
 
 const readExesVersion = () => {
   let spawn = require("child_process").spawnSync;
@@ -752,7 +480,7 @@ app.on("ready", async () => {
         label: locale.messages["appMenu.quit"],
         accelerator: "Command+Q",
         click() {
-          cleanShutdown();
+          cleanShutdown(logger, mainWindow, app, dcrdPID, dcrwPID);
         }
       } ]
     }, {
@@ -906,7 +634,7 @@ app.on("ready", async () => {
 app.on("before-quit", (event) => {
   logger.log("info","Caught before-quit. Set decredition as was closed");
   event.preventDefault();
-  cleanShutdown();
+  cleanShutdown(logger, mainWindow, app, dcrdPID, dcrwPID);
   setMustOpenForm(true);
   app.exit(0);
 });
